improvement: pipeline tpl

+++ pkg/utils/script.go
@@ -5,18 +5,18 @@ import (
 
 	"net/http"
 	"time"
+	"strings"
 )
 
 func ScriptInstallOpscli() string {
 	return fmt.Sprintf(`curl %s | sh -`, GetAvailableUrl("https://raw.githubusercontent.com/shaowenchen/opscli/main/getopscli.sh"))
 }
 
-func ScriptInstallMetricsServer() string {
+func ScriptInstallMetricsServer(clear string) string {
+	if  strings.ToLower(clear) == "true" {
+		return fmt.Sprintf(`kubectl delete -f %s`, GetAvailableUrl("https://raw.githubusercontent.com/shaowenchen/image-syncer/main/kubernetes/metrics-server-0.5.0.yaml"))
+	}
 	return fmt.Sprintf(`kubectl apply -f %s`, GetAvailableUrl("https://raw.githubusercontent.com/shaowenchen/image-syncer/main/kubernetes/metrics-server-0.5.0.yaml"))
-}
-
-func ScriptRemoveMetricsServer() string {
-	return fmt.Sprintf(`kubectl delete -f %s`, GetAvailableUrl("https://raw.githubusercontent.com/shaowenchen/image-syncer/main/kubernetes/metrics-server-0.5.0.yaml"))
 }
 
 func ScriptAddHost(ip, domain string) string {
